src/requirements.txt
dependency-injector==4.43.0
python-dotenv==1.0.1
six==1.16.0
